Add pastebody component
