+++ client/src/pages/Paste.js
@@ -2,19 +2,17 @@ import React from 'react';
 import Loading from '../components/Loading';
 import Alert from '../components/Alert';
 import styled from '@emotion/styled';
-import DateTime from '../components/DateTime';
 import useGetPaste from '../hooks/useGetPaste';
 import Button from '../components/Button';
 import FullContainer from '../components/FullContainer';
-import Author from '../components/Author';
 import PropTypes from 'prop-types';
+import PasteBody from '../components/PasteBody';
 
-const PasteArea = styled.div`
-  margin: 20px;
-`;
-
-const CreatedAt = styled(DateTime)`
-  margin: 10px;
+const WarningButton = styled(Button)`
+  background: red;
+  color: whitesmoke;
+  margin: 20px 0;
+  padding: 10px 20px;
 `;
 
 function Paste({ match }) {
@@ -31,13 +29,11 @@ function Paste({ match }) {
           <Button onClick={doGet}>Try again</Button>
         </>
       )}
-      {paste && (
+      {paste && !paste.oneTimeView && <PasteBody paste={paste} />}
+      {paste && paste.oneTimeView && (
         <>
-          <CreatedAt date={new Date(paste.createdAt)}>
-            {new Date(paste.createdAt).toDateString()}
-          </CreatedAt>
-          <Author name={paste.author} />
-          <PasteArea>{paste.value}</PasteArea>
+          <label>You can see it only once! Are you ready?</label>
+          <WarningButton>YES!!</WarningButton>
         </>
       )}
     </FullContainer>
